Отрисовка анимаций независимо от FPS

+++ src/game/common/vector.hpp
@@ -1,15 +1,15 @@
 #pragma once
 struct Vector {
-	Vector(double X, double Y) :x(X), y(Y) {};
+	Vector(float X, float Y) :x(X), y(Y) {};
 	Vector() {};
 	Vector operator+(const Vector& other) const { return Vector(x + other.x, y + other.y); };
 	Vector operator-(const Vector& other) const { return Vector(x - other.x, y - other.y); };
-	double x=0.0;
-	double y=0.0;
+	float x=0.0;
+	float y=0.0;
 };
 
 struct Quad {
-	Quad(double X, double Y, double XX, double YY){
+	Quad(float X, float Y, float XX, float YY){
 		topleft.x = X;
 		topleft.y = Y;
 		widthheight.x = XX;
